+++ src/main/java/com/airjnc/mail/service/MailService.java
@@ -1,7 +1,7 @@
 package com.airjnc.mail.service;
 
 import com.airjnc.mail.dto.request.MailSendCertificationCodeToEmailReq;
-import com.airjnc.user.dto.UserDto.UserStatus;
+import com.airjnc.user.dto.UserWhereDto.UserStatus;
 import com.airjnc.user.dto.response.UserResp;
 import com.airjnc.user.service.UserService;
 import lombok.RequiredArgsConstructor;
