+++ src/main/java/com/airjnc/user/dao/UserMapper.java
@@ -1,9 +1,8 @@
 package com.airjnc.user.dao;
 
 import com.airjnc.user.domain.UserEntity;
-import com.airjnc.user.dto.UserDto;
-import com.airjnc.user.dto.UserDto.UserStatus;
-import com.airjnc.user.dto.UserSaveDto;
+import com.airjnc.user.dto.UserWhereDto;
+import com.airjnc.user.dto.UserWhereDto.UserStatus;
 import java.util.Optional;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
@@ -11,13 +10,13 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface UserMapper {
 
-  int delete(Long id);
+  int create(UserEntity userEntity);
+
+  boolean exists(UserWhereDto userWhereDto);
 
   Optional<UserEntity> findById(@Param("id") Long userId, @Param("status") UserStatus status);
 
-  Optional<UserEntity> findByWhere(UserDto userDto);
-
-  int create(UserSaveDto userSaveDTO);
+  Optional<UserEntity> findByWhere(UserWhereDto userWhereDto);
 
   int save(UserEntity userEntity);
 }
