+++ src/main/java/com/airjnc/user/dao/UserRepository.java
@@ -1,19 +1,18 @@
 package com.airjnc.user.dao;
 
 import com.airjnc.user.domain.UserEntity;
-import com.airjnc.user.dto.UserDto;
-import com.airjnc.user.dto.UserDto.UserStatus;
-import com.airjnc.user.dto.UserSaveDto;
+import com.airjnc.user.dto.UserWhereDto;
+import com.airjnc.user.dto.UserWhereDto.UserStatus;
 
 public interface UserRepository {
 
-  void delete(Long id);
+  UserEntity create(UserEntity userEntity);
+
+  boolean exists(UserWhereDto userWhereDto);
 
   UserEntity findById(Long userId, UserStatus status);
 
-  UserEntity findByWhere(UserDto userDto);
-
-  UserEntity create(UserSaveDto userSaveDTO);
+  UserEntity findByWhere(UserWhereDto userWhereDto);
 
   UserEntity save(UserEntity userEntity);
 }
