+++ src/main/java/com/airjnc/user/service/AuthService.java
@@ -4,7 +4,7 @@ import com.airjnc.common.service.StateService;
 import com.airjnc.common.util.enumerate.SessionKey;
 import com.airjnc.user.dao.UserRepository;
 import com.airjnc.user.domain.UserEntity;
-import com.airjnc.user.dto.UserDto;
+import com.airjnc.user.dto.UserWhereDto;
 import com.airjnc.user.dto.request.AuthLogInReq;
 import com.airjnc.user.dto.response.UserResp;
 import com.airjnc.user.util.UserModelMapper;
@@ -24,7 +24,8 @@ public class AuthService {
   private final StateService stateService;
 
   public UserResp logIn(AuthLogInReq authLogInReq) {
-    UserEntity userEntity = userRepository.findByWhere(UserDto.builder().email(authLogInReq.getEmail()).build());
+    UserWhereDto userWhereDto = UserWhereDto.builder().email(authLogInReq.getEmail()).build();
+    UserEntity userEntity = userRepository.findByWhere(userWhereDto);
     userCheckService.passwordShouldBeMatch(authLogInReq.getPassword(), userEntity.getPassword());
     stateService.create(SessionKey.USER, userEntity.getId());
     return userModelMapper.userEntityToUserResp(userEntity);
